fix: product test fixes
